requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-dotenv
passlib[bcrypt]
jwt
